Reduce microframesPerTransfer from 512 to 128 in order to prevent LowLatencyCreateBuffer() from returning kIOReturnNoMemory.
This number isn't that important, but bigger is better because it reduces the chance of any playback hiccups, and memory is cheap. I'm not sure how the OS decides what the limit should be, but we might want to dynamically determine it in the future. On 10.6.5, even 129 is 1 too many.