Any time you have a loop with a boolean flag that might be set over the course of it, that's a hint that you should be building an array instead. In this case, we had two loops using that pattern, and they're much better off now.
Replaces -deviceAddedWithIterator: with -devicesWithIterator:, returning the devices themselves instead of a flag. Then the controller builds an array of all the devices. Dynamic dispatch is used to display each item, whether it's a real device or just an error. Then, if there were no devices/errors, display the "no devices found" error.