New class ECVFrameRateConverter.
New frame rate converters accept a source rate and a target rate, and then can be queried for the frame repeat count of each frame in succession in order to perform the conversion.

This is the perfect use case for unit testing, because of the complexity of the algorithm and the number of possible inputs (and results), but I still haven't really looked into testing frameworks.